Complete CRUD operations for movie entity and extend token session expiration time

+++ src/main/java/com/myproject/movie_booking/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie_booking.repositories;
 import com.myproject.movie_booking.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {}
+public interface CategoryRepository extends JpaRepository<Category, Integer> {}
